% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convexBoundedLogitCom.R
\name{convexBoundedLogitCom}
\alias{convexBoundedLogitCom}
\title{Ensemble predict function for convex combination of predictions on the logit
scale, backtransformed to original scale}
\usage{
convexBoundedLogitCom(p, alpha, trimLogit = 1e-05, lower, upper)
}
\arguments{
\item{p}{An n x K matrix of predictions from K different online learners}

\item{alpha}{A K row vector of weights}
}
\value{
prediction An n-length vector of super learner predictions
}
\description{
This function computes the ensemble prediction for the super learner
based on a convex combination of weights on the logit scale and transforms
predictions back to the original scale based on the value of \code{lower}
and \code{upper}.
}

