% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convexLogitCom.R
\name{convexLogitCom}
\alias{convexLogitCom}
\title{Ensemble predict function for convex combination of predictions}
\usage{
convexLogitCom(p, alpha, trimLogit = 1e-05)
}
\arguments{
\item{p}{An n x K matrix of predictions from K different online learners}

\item{alpha}{A K row vector of weights}
}
\value{
prediction An n-length vector of super learner predictions
}
\description{
This function computes the ensemble prediction for the super learner
based on a convex combination of weights
}

