% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialWt_Vfold.R
\name{initialWt_Vfold}
\alias{initialWt_Vfold}
\title{Compute initial estimates for online learning algorithms}
\usage{
initialWt_Vfold(Y, X, SL.library, V = 5, lossFn, ensembleControl,
  initialHessian = FALSE)
}
\arguments{
\item{Y}{The outcome vector, usually of length \code{nl}.}

\item{X}{The data.frame of predictors, with \code{nl} rows}

\item{SL.library}{The K length character vector of wrappers for candidate
online learners.}

\item{V}{The number of folds for V-fold cross validation, defaults to 5}

\item{lossFn}{The loss function used by the super learner}

\item{ensembleControl}{The list of controls for the ensemble method. The 
function uses \code{ensembleControl$ensemblePredictFn} to compute predictions
on held out data to minimize the \code{lossFn} using \code{solnp}.}

\item{hessian}{Boolean of whether to return the estimated Hessian at the 
solution of the \code{solnp} algorithm. Must be set to \code{TRUE} if 
\code{ensembleControl$updateWtFn} requires an estimate of the Hessian to 
update the weights (e.g., if \code{updateWtFn} corresponds to second-order
stochastic gradient descent).}
}
\value{
dsl The name of the algorithm from \code{SL.library} that was the
discrete super learner.

alpha A K-length vector of initial weight estimates for the super 
learner

otherWtArgs A list with entry \code{hessian} containing the estimated
hessian (set to null if \code{initialHessian = FALSE}).
}
\description{
This function creates a list of otherArgs used by each online algorithm by
calling learner.init for each learner in \code{SL.library}.
}

