% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onlinesl.R
\name{onlinesl}
\alias{onlinesl}
\title{Computes the online super learner for a data set}
\usage{
onlinesl(Y, X, family, SL.library, nl = 200, parallel = "seq",
  ensembleControl = list(initialWtFn = "initialWt_Vfold", ensemblePredictFn =
  "predict_convexLinCom", updateWtFn = "sgdWt_convexLinCom", maxCorrectWtStep =
  100), trace = list(verbose = FALSE, saveRisk = NULL), lossFn = ".mse")
}
\arguments{
\item{inputParameter1}{A description of the input parameter}

\item{inputParameter2}{A description of the input parameter}
}
\value{
output A description of the object the function outputs
}
\description{
This function uses online cross-validation to gauge the performance of 
a user-select library of online algorithms. More description to come.
}
\examples{
Examples to come

}
\keyword{keywords}

