% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SL.sgd.boundedLogistic-predict-method.R
\name{predict.SL.sgd.boundedLogistic}
\alias{predict.SL.sgd.boundedLogistic}
\title{Computes a prediction for sgd poisson regression}
\usage{
\method{predict}{SL.sgd.boundedLogistic}(object, newdata, ...)
}
\arguments{
\item{newdata}{A \code{data.frame} object with data on which predictions are 
desired.}

\item{fit}{The current object fit of class \code{SL.sgd.poisson}}
}
\value{
pred A vector of predictions for current sgd poisson regression.
}
\description{
Get a prediction on a new data point for the current sgd poisson regression 
fit.
}
\examples{
Examples to come

}

