% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convexLinCom.R, R/predict_convexLinCom.R
\name{predict_convexLinCom}
\alias{predict_convexLinCom}
\title{Ensemble predict function for convex combination of predictions}
\usage{
predict_convexLinCom(p, alpha)

predict_convexLinCom(p, alpha)
}
\arguments{
\item{p}{An n x K matrix of predictions from K different online learners}

\item{alpha}{A K row vector of weights}

\item{p}{An n x K matrix of predictions from K different online learners}

\item{alpha}{A K row vector of weights}
}
\value{
prediction An n-length vector of super learner predictions

prediction An n-length vector of super learner predictions
}
\description{
This function computes the ensemble prediction for the super learner
based on a convex combination of weights

This function computes the ensemble prediction for the super learner
based on a convex combination of weights
}

