% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onlinesl-print-method.R
\name{print.onlinesl}
\alias{print.onlinesl}
\title{Print the output of onlinesl object}
\usage{
\method{print}{onlinesl}(object, ...)
}
\arguments{
\item{object}{An object of class \code{onlinesl}}
}
\description{
Print the output of onlinesl object
}

