% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgdWt_convexLinCom.R
\name{sgdWt_convexLinCom}
\alias{sgdWt_convexLinCom}
\title{Perform first order stochastic gradient descent update of super learner 
weights}
\usage{
sgdWt_convexLinCom(Y, slFit.t, p.t, tplus1, stepSize = NULL)
}
\arguments{
\item{Y}{The outcome at iteration t}

\item{slFit.t}{A named list with a component named alpha.t that contains the
1-column matrix of current estimate of the super learner weights}

\item{p.t}{The predictions from the various online algorithms at time t}

\item{tplus1}{The iteration of the online algorithm}

\item{stepSize}{The size of the step to take in the direction of the
gradient. If \code{stepSize=NULL} (default) the function uses 
\code{1/tplus1}.}
}
\value{
alpha A matrix of updated weights.
}
\description{
This function performs a single step of gradient descent on the weight 
vector for the super learner weights and projects the resulting vector onto
the L1-simplex via the internal function .projToL1Simp. The function returns
the updated weight vector.
}

