% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SL.sgd.poisson.R
\name{SL.sgd.poisson}
\alias{SL.sgd.poisson}
\title{Computes a stochastic gradient step for poisson regression}
\usage{
SL.sgd.poisson(Y, X, newX, t, fit, initial = FALSE, formula = "Y ~ .",
  stepSize = function(t) {     1/t }, ...)
}
\arguments{
\item{Y}{The numeric input outcome}

\item{X}{The data.frame of predictors (ignored if \code{initial = FALSE})}

\item{newX}{The data.frame to return predictions for.}

\item{t}{The numeric step number (ignored if \code{initial = TRUE})}

\item{fit}{A list of the current fit with named entry theta corresponding 
to the regression parameter estimates at the previous iteration and formula
corresponding to the regression formula used by the function.}

\item{initial}{Boolean as to whether the function should be called to obtain
initial regression parameter estimates or whether it should perform an SGD
updating step}

\item{formula}{The regression formula to be used, specified as a character}

\item{stepSize}{A function evaluating what step size to use in the SGD 
algorithm (ignored if \code{initial = TRUE})}

\item{initial}{A boolean indicating whether to execute the function for
initial parameter estimates or for an sgd update step.}
}
\value{
fit An object of class \code{SL.sgd.poisson}. A named list with 
entries \code{theta} and \code{formula} consisting of the current regression 
parameter estimates and the regression formula.

pred Predictions on newX either based on an initial call to \code{glm}
(if \code{initial = TRUE}) or based on an SGD update (if \code{initial=FALSE})
}
\description{
Wrapper function for using Poisson GLM with online SL. If the function is 
called with \code{initial=TRUE} then fits a Poisson regression using the 
specified formula and returns predicted values on data frame \code{newX}. Any 
input for \code{fit} is ignored. This is called by \code{onlinesl} to obtain
starting values for the SGD algorithm. If the function is called with 
\code{inital=FALSE}, the function computes a prediction on \code{newX} based 
on the current \code{fit} and updates the value of \code{fit} by taking a  
step in the direction of the gradient. The function returns a list with named
entries \code{fit} and \code{pred}. \code{fit} contains the current regression
parameter estimate, while \code{pred} contains the predictions on the 
\code{newX} data frame.
}
\examples{
Examples to come

}

